#1 First pass at README documentation
This patch includes a first pass at writing up our README.

Unfortunately, there's currently not a whole lot fleshed out, so there's not a whole lot that's trivial to document.

This patch is in service of #1, which is to trivially document already fleshed out plans.

Engaging in this activity has luckily discovered several other tickets that will further flesh out implementation details and future work.

Some of those tickets include #2, #3, #4, #5, etc.

This patch merely documents the technologies we intend to use as well as some documentation that has been written up (also in service of #1) on the wiki.