#6 Initial commit of skeleton code for our front end
This patch adds in a bunch of skeleton code on which we intend to build our front end.

It was created via create-react-app (see https://github.com/facebook/create-react-app).

We're current missing a favicon, but that'll be handled in #7.

There's pretty much nothing there.

This patch simply adds an almost flat directory heirarchy for our components.