#14 Initial commit of testing framework.
This necessitates patching in the current progress of #6. We have to patch in the current functinoality that's supported so that the tests can hit it. This'll put the project in a somewhat broken state, but the functionality was already broken or non-existant.

Thus, this patch contains a significant amount of work on the endpoints we plan on supporting via Firebase.

See our tests in the swing_dance_scores/test directory for documentation on how the endpoints work.

The tests we have so far cover the basic functionality for all the endpoints as well as authentication tests for all the end points.

We have two integration-style end-to-end tests (see testAllWebServicesEndToEnd and testAllWebServicesWrtAuthentication). All the supported endpoints are documented there.

There's almost a one-to-one correspondence between our JS functions and our tested end points. I'll show them below:
                     getAllCompetitions                     /competitions                 GET                   No Authentication Required
		     postOneCompetition                     /createCompetition            POST                  Authentication Required
		     findCompetition                        /competition/:competitionId   GET                   No Authentication Required
		     deleteCompetition                      /competition/:competitionId   DELETE                Authentication Required
		     signup                                 /signup                       POST                  No Authentication Required
		     login                                  /login                        POST                  No Authentication Required
		     uploadImage                            /users/image                  POST                  Authentication Required
		     addUserDetails                         /users/addUserDetails         POST                  Authentication Required
		     getUserData                            /users/getUserDetails         GET                   Authentication Required

This patch also includes some changes to the top-level CLI. We renamed some things. The most important addition was the addition of a top-level CLI command to run all tests. This'll make development faster.

This patch also includes DB-side trigger to create notifications when competitions are deleted. We'll need to add a test for this as well as other triggers.

We support uploading images! We unfortunately have yet to have developed a test for that. Stay tuned.

There's still some work to be done as we need to add a test for the "findCompetition" endpoint.

This is a progress patch.