#6 Rename "front_end" directory to "back_end"
We realized that all the functionality we've been putting into our front_end directory was actually for the backend as all of that code supports our endpoints.

We'd like to have a separate directory for our actual front end to be separate from the back end code that supports out endpoints.

The back_end directory should contain all the code for Firebase and the webservices we support.

The front_end directory should contain all of our React code and front end code that utilize the webservices we support.

We changed our driver script at swing_dance_scores.py in according with this renaming.