#20 Unstub _start_development_servers
This commit provides the initial support for our top-level CLI option -start-development-servers that'll start the front and back end servers.

Additionally, this patch provides support for some convenient sanity checking of background processes during testing.

Our testing relies on using a fixed set of ports.

Sometimes, when tests go badly, there are lingering processes hogging the desired ports.

Running the tests from the CLI will now ask if we want to have those processes killed prior to running the tests.

This is Linux only, so we might have to featurize this depending on the platform.