#20 Make -start-development-servers not leave around lingering processes.
There was a bug where we were leaving around lingering processes when running

             ./swing_dance_scores.py -start-development-servers

This stemmed from the fact that we only shut down the front and back end servers when there was an error or when there was a keyboard interrupt.

On normal execution, nothing would happen.

We now wait around until the user interrupts. This'll give a convenient way to shut down the servers.

There was also the problem that the -start-development-servers functionality was still entirely stubbed out via a NotImplementedError exception. That's been removed.