#24 Initial commit of front end login page and sign up page
This patch includes the initial commit of the front end login page and sign up page.

This patch inludes several other changes as well.

We've had to update our Firebase backend (these changes have been deployed) so that we accept CORS. We encountered quite a bit of trouble with this when trying to have our front end on one server and our back end on another (this is the eventual plan for deployment as we plan for our Firebase backend to purely serve as an API). We have plans in #24 to write a test ensuring we don't have these woes again. This necessitated an NPM library instal and use of it in our backend's index.js file.

We also as part of #22 in pass modified our signup and login API functions to include a refresh token.

This patch includes a lot of infrastructure for our front end.

The homepage via HomePage.js is just a stub right now. That's next up on the list to get up and working.

Our login and sign up pages use Material UI. It's not clear if we're going to stick with that in the long run or not. There are still a lot of aesthetic changes that we plan on making.

As part of our front end we have a top navigation bar that we plan on making part of every page (with possibly different options). Planning is still happening in that respect.

This patch also includes a basic use of redux to store our global state. See the redux directory. Our reducers are not yet complex enough to warrant their own file. We'll get there soon. We're currently using redux to track our authentication status.

This patch also includes some aesthetic text changes to our CLI's prints.