#24 Progress patch on maintaining authentication
This is a progress patch on the implementation of our front end for a login and sign up page.

We want to add these in order to make it simpler for us to manually test our Redux capabilities prior to writing unit tests to be run automatically.

This patch makes hte login page work.

The signup page is still not working.

We have a functional log out button.

We have stub pages for the "My Competitions" page and the "My Profile" pages.

This patch includes a lot of work in noting authentication issues, sign up issues, login issues, etc. in our global state via Redux.