#28 Verify production DB is free of automatically generated test data
We've previously had problems with DB administration with all the test users being generated and placed into our production DB.

This is because we currently do not have a way of testing authentication solely via the firestore emulator.

This patch makes it so that test clear out any test users from the production DB.

We don't modify any other tables in the production DB, so no further production DB pollution needs to be accounted for that is known.

The changes in this patch require authenticating the test runs so that we can change the production DB.

This requires us to have a serviceAccountKey.json ; see https://console.firebase.google.com/u/0/project/swing-dance-scores/settings/serviceaccounts/adminsdk and https://firebase.google.com/docs/admin/setup for more information.

We don't want this information stored in the repository, so we added .gitignore files to make sure that they aren't suggested so that we don't add them accidentally.

We also added some automatically generated .gitignore files created by the firebase utilities.

We also added some corresponding .gitignore files for the front end React code.

This patch also adds functionality for a delete user endpoint. We'll add test cases for those in a future patch.

This patch also adds a sweep test that verifies that the production DB is free of test users.

This'll lead to future gains as we have re-usable infrastructure for production DB data validation.

This patch also includes some convenience functionality that'll clear the DB of test users prior to running tests.