#35 Initial commit of testSignUpWebserviceViaLocalFirestoreEmulator that verifies the "Create User" endpoint works as intended
This patch includes the test testSignUpWebserviceViaLocalFirestoreEmulator in the file test_sign_up_webservice_via_local_firestore_emulator.py that makes sure that we can simply sign up a new user. It does not hit the production dB (outside of using it for authentication). It's intended to help isolate functionality-specific failures. The test ensures the deletion of any test users stored in the production DB that were created out of a need to for authentication (that's dealt with remotely on the production DB).