#35 Update incorrect test expectations
While working on the newer test testSignUpWebserviceViaLocalFirestoreEmulator, we noticed a few bugs.

We noticed that in the test testAllWebServicesEndToEndViaLocalFireStoreEmulator, we had an incorrect message associated with a test expectation. This caused us to waste a lot of time debugging something that wasn't actually a problem . The problem was that we were printing out the wrong URI (the "get all competitions" endpoint uri) when the teset was verifying results from the "sign up" or "create user" endpoint.

This was also a problem in testSignUpWebserviceViaLocalFirestoreEmulator as it was based on testAllWebServicesEndToEndViaLocalFireStoreEmulator.

We also updated several typos, c&p issues, etc. in testSignUpWebserviceViaLocalFirestoreEmulator.

While debugging testSignUpWebserviceViaLocalFirestoreEmulator, we added some more useful misc utilities such as extract_lines_from_subprocess and debug_log.