#35 Partial progress on fixing testAllWebServicesWrtAuthenticationViaLocalFireStoreEmulator
This is a progress patch that fixes a significant portion of testAllWebServicesWrtAuthenticationViaLocalFireStoreEmulator.

findCompetitionsModifiableByUser didn't handle the case where no authentication information was provided. This is problematic since we want all public competitions to be shown in this case.

We refactored visibleCompetitionInfosForUserWithHandle to be implemented within visibleCompetitionInfos since it didn't make sense to have this abstraction as information regarding the existence of the handle was necessary for visibleCompetitionInfos to work properly. Passing around handles that could possibly be NULL seemed not particularly future-proof. However, we did maintain a similar abstraction by creating publicCompetitionInfos.

possiblyFBAuthenticate had a huge bug in that it didn't handle the case where authorization information wasn't provided. It simply did nothing, which caused hanging.

We did some refactoring of testAllWebServicesWrtAuthenticationViaLocalFireStoreEmulator along the way to update test expectations and also provide useful STDOUT information we could gather from the firebase emulator process.

Also, findCompetitionById had a lingering debug print that we removed.

Also, we fixed a logging issue in createNotificationOnCompetitionDelete and FBAuthenticate.