#36 Make testAllWebServicesWrtAuthenticationViaLocalFireStoreEmulator pass
This patch makes testAllWebServicesWrtAuthenticationViaLocalFireStoreEmulator pass.

The main changes relevant to this commit pertain to test expectations.

testAllWebServicesWrtAuthenticationViaLocalFireStoreEmulator tests competition creation functionality currently with only private competitions (there's a @todo to handle public competitions). The headers were not set correctly when the test tried to hit the relevant endpoints. This commit addresses this problem.