#36 Make testAllWebServicesEndToEndViaDeployedApp pass
This patch completes the task of #36 as it makes all tests pass.

The implementation of testAllWebServicesEndToEndViaDeployedApp is not yet complete, so there's more work to do on that front.

There was an issue in the implementation of possiblyFBAuthenticate. It worked on local firestore emulator but would not deploy properly since there was a missing "return" call. The problem was mostly that deployment enforced betterpractices by requiring a "return" call.

This patch also partially implements the top-level CLI command "-deploy" and unstubs the backend firebase deployment functionality.

testAllWebServicesEndToEndViaDeployedApp only had minor issues, e.g. typos in variable names. Most issues were with deployment.